Rename module ogn.exceptions to ogn.parser.exceptions
