+++ ogn/gateway/client.py
@@ -4,7 +4,7 @@ from time import time
 
 from ogn.gateway import settings
 from ogn.aprs_parser import parse_aprs
-from ogn.exceptions import AprsParseError, OgnParseError, AmbigousTimeError
+from ogn.parser.exceptions import AprsParseError, OgnParseError, AmbigousTimeError
 
 
 def create_aprs_login(user_name, pass_code, app_name, app_version, aprs_filter=None):
