+++ ogn/parser/exceptions.py
@@ -9,17 +9,16 @@ class AprsParseError(Exception):
     def __init__(self, aprs_string):
         self.aprs_string = aprs_string
 
-        self.message = "This is not a valid APRS string: {}".format(aprs_string)
+        self.message = "This is not a valid APRS packet: {}".format(aprs_string)
         super(AprsParseError, self).__init__(self.message)
 
 
 class OgnParseError(Exception):
-    """Parse error while parsing an aprs packet substring."""
-    def __init__(self, substring, expected_type):
-        self.substring = substring
-        self.expected_type = expected_type
+    """Parse error while parsing an ogn message from aprs comment."""
+    def __init__(self, aprs_comment):
+        self.aprs_comment = aprs_comment
 
-        self.message = "For type {} this is not a valid token: {}".format(expected_type, substring)
+        self.message = "This is not a valid OGN message: {}".format(aprs_comment)
         super(OgnParseError, self).__init__(self.message)
